+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 Any notable changes to this project will be documented in this file.
 
 ## 0.5.3
+
 ### Feature:
 - [EKNotificationMessage has broken layout #64](https://github.com/huri000/SwiftEntryKit/issues/64) - Add margins to `EKNotificationMessage`.